#22 - Divided Chat window into smaller components
