+++ src/components/Chat/MessageList/styles.scss
@@ -1,0 +1,16 @@
+.pos-absolute {
+    position: absolute !important;
+  }
+
+.chat-messages-box {
+    display: flex;
+    flex-flow: column nowrap;
+    align-items: left;
+    justify-content: top;
+    top: 10em;
+    left: 0;
+    width: 100%;
+    height: 50vh;
+    overflow-y: auto;
+    margin: 0 auto;
+}
