+++ src/components/Chat/UserInput/styles.scss
@@ -1,0 +1,10 @@
+.pos-absolute {
+    position: absolute !important;
+  }
+
+.chat-input-field {
+    bottom: 5em;
+    left: 0;
+    width: 100%;
+    margin: 0 auto;
+  }
